Clean up test code + imports for Java 16

+++ common/src/main/java/io/github/soir20/moremcmeta/client/texture/AnimationComponent.java
@@ -76,7 +76,7 @@ public class AnimationComponent implements ITextureComponent {
 
                     if (timeOptional.isPresent()) {
                         long currentTime = timeOptional.get();
-                        int ticksToAdd = (int) Math.floorMod(currentTime - ticks, SYNC_TICKS);
+                        int ticksToAdd = Math.floorMod(currentTime - ticks, SYNC_TICKS);
 
                         ticks += ticksToAdd;
                         FRAME_MANAGER.tick(ticksToAdd);
